start up helper
