+++ internal/cli/cliHandlers/startHandler.go
@@ -1,14 +1,11 @@
 package cliHandlers
 
 import (
-	"fmt"
 	"os"
 	"strings"
 
+	wdb "github.com/TanmoySG/wunderDB/cmd/entrypoint"
 	"github.com/TanmoySG/wunderDB/internal/config"
-	"github.com/TanmoySG/wunderDB/internal/server/lifecycle/shutdown"
-	"github.com/TanmoySG/wunderDB/internal/server/lifecycle/startup"
-	"github.com/charmbracelet/lipgloss"
 	"github.com/urfave/cli/v2"
 )
 
@@ -43,24 +40,9 @@ func (ch cliHandler) StartOptHandler(ctx *cli.Context) error {
 
 	setEnv(ctx)
 
-	var style = lipgloss.NewStyle().
-		Bold(true).
-		Foreground(lipgloss.Color("#7D56F4")).PaddingTop(1)
-
-	fmt.Println(style.Render("Starting wunderDb..."))
-
-	c, err := config.Load()
-	if err != nil {
-		return fmt.Errorf("error loading configurations: %s", err)
+	if err := wdb.EntryPoint(); err != nil {
+		return err
 	}
-
-	w, err := startup.Prepare(*c)
-	if err != nil {
-		return fmt.Errorf("error starting wdb server: %s", err)
-	}
-
-	shutdown.Listen(*w, *c) // listens to shutdown signals
-	startup.Start(w, c)     // starts server and initial setup
 
 	return nil
 }
